refactored token management to be fully based in the tokens module
